requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
email-validator>=2.0.0
Flask-SQLAlchemy>=3.0.0
Werkzeug>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
Pillow>=9.0.0
dashscope~=1.25.5
WTForms~=3.2.1